Canal @DevDojoBrasil - Operadores pt 02 -- Relacionais

+++ @DevDojoBrasil/Maratona Java Virado no Jiraya/15 - Operadores pt 01 - Aritiméticos/maratona-java15/src/academy/devdojo/maratonajava/introducao/Aula04Operadores.java
@@ -5,7 +5,7 @@ public class Aula04Operadores {
         // + - / *
         int numero1 = 10;
         int numero2 = 20;
-        double resultado = numero1 / (double) numero2;
+        double resultado = numero1 / numero2;
         System.out.println(resultado);
     }
 }
